+++ packages/client/.gitignore
@@ -27,7 +27,9 @@ yarn-error.log*
 
 # local env files
 .env*.local
-
+.env.development
+.env.production
+.env.qa
 # vercel
 .vercel
