+++ packages/client/package.json
@@ -24,6 +24,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "4.9.4"
+    
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
@@ -36,5 +37,6 @@
     "postcss": "^8.4.21",
     "prettier": "^2.8.2",
     "tailwindcss": "^3.2.4"
+    
   }
 }
